# shared/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "DialogueResponse": {
        "additionalProperties": false,
        "properties": {
          "character_id": {
            "title": "Character Id",
            "type": "string"
          },
          "character_name": {
            "title": "Character Name",
            "type": "string"
          },
          "role": {
            "enum": [
              "companion",
              "npc"
            ],
            "title": "Role",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "character_id",
          "character_name",
          "role",
          "text"
        ],
        "title": "DialogueResponse",
        "type": "object"
      },
      "ErrorCode": {
        "enum": [
          "AUTH_CONFIGURATION_ERROR",
          "INVALID_INPUT",
          "HTTP_ERROR",
          "NOT_FOUND",
          "CONFLICT",
          "UNAUTHORIZED",
          "FORBIDDEN",
          "INTERNAL_ERROR"
        ],
        "title": "ErrorCode",
        "type": "string"
      },
      "ErrorDetail": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorCode"
              },
              {
                "type": "string"
              }
            ],
            "title": "Code"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "GameIslandState": {
        "additionalProperties": false,
        "properties": {
          "island_id": {
            "const": "joy_meadow",
            "title": "Island Id",
            "type": "string"
          },
          "name": {
            "const": "Joy Meadow",
            "title": "Name",
            "type": "string"
          },
          "restoration_level": {
            "maximum": 100,
            "minimum": 0,
            "title": "Restoration Level",
            "type": "integer"
          },
          "restored": {
            "title": "Restored",
            "type": "boolean"
          }
        },
        "required": [
          "island_id",
          "name",
          "restoration_level",
          "restored"
        ],
        "title": "GameIslandState",
        "type": "object"
      },
...
      "GameStateResponse": {
        "additionalProperties": false,
        "properties": {
          "dialogue": {
            "items": {
              "$ref": "#/components/schemas/DialogueResponse"
            },
            "title": "Dialogue",
            "type": "array"
          },
          "inventory": {
            "items": {
              "$ref": "#/components/schemas/InventoryItemResponse"
            },
            "title": "Inventory",
            "type": "array"
          },
          "island": {
            "$ref": "#/components/schemas/GameIslandState"
          },
          "journal": {
            "items": {
              "$ref": "#/components/schemas/JournalEntryResponse"
            },
            "title": "Journal",
            "type": "array"
          },
          "quest": {
            "$ref": "#/components/schemas/QuestStateResponse"
          },
          "recipe": {
            "$ref": "#/components/schemas/RecipeStateResponse"
          },
          "save": {
            "$ref": "#/components/schemas/SaveStatusResponse"
          },
          "started": {
            "title": "Started",
            "type": "boolean"
          }
        },
        "required": [
          "started",
          "island",
          "dialogue",
          "inventory",
          "quest",
          "recipe",
          "journal",
          "save"
        ],
        "title": "GameStateResponse",
        "type": "object"
      },
...
      "InventoryItemResponse": {
        "additionalProperties": false,
        "properties": {
          "collected": {
            "title": "Collected",
            "type": "boolean"
          },
          "ingredient_id": {
            "enum": [
              "vanilla_orchid",
              "honey_bloom"
            ],
            "title": "Ingredient Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "quantity": {
            "minimum": 0,
            "title": "Quantity",
            "type": "integer"
          }
        },
        "required": [
          "ingredient_id",
          "name",
          "quantity",
          "collected"
        ],
        "title": "InventoryItemResponse",
        "type": "object"
      },
...
      "JournalCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "island_id": {
            "const": "joy_meadow",
            "title": "Island Id",
            "type": "string"
          }
        },
        "required": [
          "island_id"
        ],
        "title": "JournalCreateRequest",
        "type": "object"
      },
      "JournalEntryResponse": {
        "additionalProperties": false,
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "emotion": {
            "title": "Emotion",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "recipe_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recipe Name"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "content",
          "emotion",
          "recipe_name",
          "created_at"
        ],
        "title": "JournalEntryResponse",
        "type": "object"
      },
      "JournalResponse": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/JournalEntryResponse"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "JournalResponse",
        "type": "object"
      },
...
      "QuestCompleteRequest": {
        "additionalProperties": false,
        "properties": {
          "quest_id": {
            "const": "joy_first_recipe",
            "title": "Quest Id",
            "type": "string"
          }
        },
        "required": [
          "quest_id"
        ],
        "title": "QuestCompleteRequest",
        "type": "object"
      },
      "QuestProgressRequest": {
        "additionalProperties": false,
        "properties": {
          "quest_id": {
            "const": "joy_first_recipe",
            "title": "Quest Id",
            "type": "string"
          }
        },
        "required": [
          "quest_id"
        ],
        "title": "QuestProgressRequest",
        "type": "object"
      },
      "QuestStartRequest": {
        "additionalProperties": false,
        "properties": {
          "quest_id": {
            "const": "joy_first_recipe",
            "title": "Quest Id",
            "type": "string"
          }
        },
        "required": [
          "quest_id"
        ],
        "title": "QuestStartRequest",
        "type": "object"
      },
      "QuestStateResponse": {
        "additionalProperties": false,
        "properties": {
          "can_complete": {
            "title": "Can Complete",
            "type": "boolean"
          },
          "collected_ingredients": {
            "items": {
              "type": "string"
            },
            "title": "Collected Ingredients",
            "type": "array"
          },
          "crafted": {
            "title": "Crafted",
            "type": "boolean"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "quest_id": {
            "title": "Quest Id",
            "type": "string"
          },
          "recipe_id": {
            "title": "Recipe Id",
            "type": "string"
          },
          "required_ingredients": {
            "items": {
              "type": "string"
            },
            "title": "Required Ingredients",
            "type": "array"
          },
          "status": {
            "enum": [
              "not_started",
              "active",
              "completed"
            ],
            "title": "Status",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "quest_id",
          "title",
          "description",
          "status",
          "required_ingredients",
          "collected_ingredients",
          "recipe_id",
          "crafted",
          "can_complete"
        ],
        "title": "QuestStateResponse",
        "type": "object"
      },
      "QuestsResponse": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/QuestStateResponse"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "QuestsResponse",
        "type": "object"
      },
      "RecipeCraftRequest": {
        "additionalProperties": false,
        "properties": {
          "recipe_id": {
            "const": "golden_vanilla_bloom",
            "title": "Recipe Id",
            "type": "string"
          }
        },
        "required": [
          "recipe_id"
        ],
        "title": "RecipeCraftRequest",
        "type": "object"
      },
      "RecipeStateResponse": {
        "additionalProperties": false,
        "properties": {
          "ability": {
            "title": "Ability",
            "type": "string"
          },
          "can_craft": {
            "title": "Can Craft",
            "type": "boolean"
          },
          "crafted": {
            "title": "Crafted",
            "type": "boolean"
          },
          "emotion": {
            "title": "Emotion",
            "type": "string"
          },
          "lore": {
            "title": "Lore",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "recipe_id": {
            "title": "Recipe Id",
            "type": "string"
          },
          "required_ingredients": {
            "items": {
              "type": "string"
            },
            "title": "Required Ingredients",
            "type": "array"
          }
        },
        "required": [
          "recipe_id",
          "name",
          "emotion",
          "lore",
          "ability",
          "required_ingredients",
          "crafted",
          "can_craft"
        ],
        "title": "RecipeStateResponse",
        "type": "object"
      },
      "RecipesResponse": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RecipeStateResponse"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "RecipesResponse",
        "type": "object"
      },
...
      "SaveStatusResponse": {
        "additionalProperties": false,
        "properties": {
          "last_event": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Event"
          },
          "last_saved_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Saved At"
          },
          "status": {
            "enum": [
              "not_saved",
              "saved"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "last_saved_at",
          "last_event"
        ],
        "title": "SaveStatusResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Flavor Odyssey Backend",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/journal": {
      "get": {
        "operationId": "get_journal",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JournalResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Journal",
        "tags": [
          "journal"
        ]
      }
    },
    "/api/journal/create": {
      "post": {
        "operationId": "create_journal_entry",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JournalCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameStateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Entry",
        "tags": [
          "journal"
        ]
      }
    },
    "/api/quests": {
      "get": {
        "operationId": "get_quests",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Quests",
        "tags": [
          "quests"
        ]
      }
    },
    "/api/quests/complete": {
      "post": {
        "operationId": "complete_quest",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestCompleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameStateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Complete Quest",
        "tags": [
          "quests"
        ]
      }
    },
    "/api/quests/progress": {
      "post": {
        "operationId": "update_quest_progress",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestProgressRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameStateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Progress",
        "tags": [
          "quests"
        ]
      }
    },
    "/api/quests/start": {
      "post": {
        "operationId": "start_quest",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestStartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameStateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Start Quest",
        "tags": [
          "quests"
        ]
      }
    },
    "/api/recipes": {
      "get": {
        "operationId": "get_recipes",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Recipes",
        "tags": [
          "recipes"
        ]
      }
    },
    "/api/recipes/craft": {
      "post": {
        "operationId": "craft_recipe",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecipeCraftRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameStateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Craft Recipe",
        "tags": [
          "recipes"
        ]
      }
    },
...
  },
...
}